Added final change?

+++ AIFG-3D-Connect-4/Grid.h
@@ -46,10 +46,10 @@ private:
 	const float m_PIECE_SIZE = 64;
 
 	// The spacing between each grid.
-	const int m_Y_SPACING_Multiplier = 50;
+	const float m_Y_SPACING_Multiplier = 50;
 
 	// The offset from the top of a grid.
-	const int m_Y_OFFSET = 15;
+	const float m_Y_OFFSET = 15;
 
 	float m_topLeftX;
